requirements.txt
google-cloud-bigquery~=3.4.0
googlemaps~=4.7.3
requests~=2.28.1
shapely~=1.8.5
